+++ apps/files_sharing/lib/Config/ConfigLexicon.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-namespace OCA\Files_Sharing\AppInfo;
+namespace OCA\Files_Sharing\Config;
 
 use NCU\Config\Lexicon\ConfigLexiconEntry;
 use NCU\Config\Lexicon\ConfigLexiconStrictness;
